pip_requirements.txt
aiogram==2.4
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.7.3
python-telegram-bot-pagination==0.0.2
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
yarl==1.4.2
